Move SimpleIni.h to fdbclient
