+++ fdbclient/CMakeLists.txt
@@ -47,6 +47,7 @@ set(FDBCLIENT_SRCS
   ConfigKnobs.h
   ConfigTransactionInterface.cpp
   ConfigTransactionInterface.h
+  ConvertUTF.h
   CoordinationInterface.h
   DatabaseBackupAgent.actor.cpp
   DatabaseConfiguration.cpp
@@ -120,6 +121,7 @@ set(FDBCLIENT_SRCS
   ServerKnobs.cpp
   ServerKnobs.h
   SimpleConfigTransaction.h
+  SimpleIni.h
   SnapshotCache.h
   SpecialKeySpace.actor.cpp
   SpecialKeySpace.actor.h
