+++ fdbclient/ConvertUTF.h
